+++ src/export.ts
@@ -1,3 +1,4 @@
 ﻿export { parse } from './renderer/parser';
 export { bind } from './renderer/bind';
 export { render } from './renderer/render';
+export { caches } from './parser/cache';