+++ src/parser/cache.ts
@@ -1,0 +1,7 @@
+import { cache as image } from './inline/media';
+import { cache as math } from './inline/math';
+
+export const caches = {
+  image,
+  math,
+};