+++ index.d.ts
@@ -10,19 +10,20 @@ export function bind(el: HTMLElement): (source: string) => HTMLElement[];
 export function render(el: HTMLElement, opts?: RenderingOptions): void;
 
 export interface RenderingOptions {
-  code?: boolean;
-  math?: boolean;
+  code?: (source: HTMLElement) => void;
+  math?: (source: HTMLElement) => void;
   media?: {
-    twitter?: boolean;
-    youtube?: boolean;
-    gist?: boolean;
-    slideshare?: boolean;
-    pdf?: boolean;
+    twitter?: (url: string) => HTMLElement | void;
+    youtube?: (url: string) => HTMLElement | void;
+    gist?: (url: string) => HTMLElement | void;
+    slideshare?: (url: string) => HTMLElement | void;
+    pdf?: (url: string) => HTMLElement | void;
+    image?: (source: HTMLElement) => HTMLImageElement | HTMLAnchorElement;
   };
 }
 
 import { Cache } from 'spica/cache';
 export const caches: {
-  readonly image: Cache<string, HTMLImageElement>;
   readonly math: Cache<string, HTMLElement>;
+  readonly image: Cache<string, HTMLImageElement>;
 };