+++ src/renderer/render/code.ts
@@ -1,4 +1,4 @@
 ﻿export function code(source: HTMLElement): void {
-  if (source.children.length > 0) return;
+  assert(source.children.length === 0);
   void Prism.highlightElement(source, false);
 }