+++ src/renderer/render/media/image.test.ts
@@ -4,9 +4,10 @@ import DOM from 'typed-dom';
 describe('Unit: renderer/render/media/image', () => {
   describe('image', () => {
     it('result', () => {
-      assert(image(DOM.img({ 'data-src': '/' }, '').element).getAttribute('src') === '/');
-      assert(image(DOM.img({ 'data-src': '/' }, '').element) instanceof HTMLImageElement);
-      assert(image(DOM.a({ img: DOM.img({ 'data-src': '/' }, '') }).children.img.element) instanceof HTMLAnchorElement);
+      assert(image(DOM.img({ 'data-src': '/', alt: '' }, '').element) instanceof HTMLImageElement);
+      assert(image(DOM.img({ 'data-src': '/', alt: '' }, '').element).getAttribute('src') === '/');
+      assert(image(DOM.img({ 'data-src': '/', alt: '' }, '').element).getAttribute('alt') === '');
+      assert(image(DOM.a({ img: DOM.img({ 'data-src': '/', alt: '' }, '') }).children.img.element) instanceof HTMLAnchorElement);
     });
 
   });