+++ src/renderer/render/media/image.ts
@@ -1,16 +1,21 @@
 ﻿import DOM from 'typed-dom';
 import { cache } from '../../../parser/inline/media';
 
-export function image(source: HTMLImageElement): HTMLElement {
+export function image(source: HTMLImageElement): HTMLImageElement | HTMLAnchorElement {
   assert(source.hasAttribute('data-src'));
-  assert(source.getAttribute('data-src') !== '');
+  assert(source.hasAttribute('alt'));
   const url = source.getAttribute('data-src')!;
-  void source.removeAttribute('data-src');
-  source = DOM.img({
-    class: 'media',
-    src: url,
-    style: 'max-width: 100%;',
-  }, () => source).element;
-  void cache.set(url, <HTMLImageElement>source.cloneNode(true));
-  return <HTMLElement>source.closest('a') || source;
+  const el = cache.has(url)
+    ? <HTMLImageElement>cache.get(url)!.cloneNode(true)
+    : cache.set(url, <HTMLImageElement>DOM.img({
+        class: 'media',
+        src: url,
+        alt: source.getAttribute('alt')!,
+        style: 'max-width: 100%;',
+      }).element.cloneNode(true));
+  return source.parentElement
+      && source.parentElement.matches('a')
+      && source.parentElement.replaceChild(el, source)
+    ? <HTMLAnchorElement>el.parentElement!
+    : el;
 }